Add a task to blink the led for wifi status.
Slow blink when not connected and then on while sending and off while doing other things.

The solution with two bits, one for send and one for done sending is needed since
there is no way to wait for a bit to be cleared (except polling).
We need to wait with setting the sent bit in the sending thread until we are all done
i.e closing any sockets or similar. Otherwise the wait function in the led task will wait indefinitely